Changed abs(0) back to abs on line 60/61
This change had caused fade time to fail in some cases. The change to abs(0) was to solve an issue a user reported when a null target value was provided. I assumed abs(0) worked the same way float(0) worked to default a 0 value when input was null, which doesn't appear to be the case. I'll investigate other solutions.